messaging_service/requirements.txt
flask
flask-socketio
python-socketio
python-engineio
eventlet
pyjwt
requests